refactor: New scanner view
